Update with test code to show the buttons off in highlight state.
We now have nicer test code that shows off the buttons in highlighted
state. They are also labelled so show what flags are present on the
buttons.

The code has also been updated so that it includes the github workflow
and the build service build (even though on my system it will still
build through the gitlab-ci build using the native build tools).